[REFACTORING] 항목명 수정
