+++ app/src/main/res/layout/activity_sign_in.xml
@@ -4,7 +4,6 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#212832"
-
     android:orientation="vertical"
     android:padding="0dp">
 
@@ -12,9 +11,8 @@
         android:id="@+id/login_glide"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        app:layout_constraintTop_toTopOf="parent"
-        android:scaleType="centerCrop"/>
-
+        android:scaleType="centerCrop"
+        app:layout_constraintTop_toTopOf="parent" />
 
     <ScrollView
         android:layout_width="match_parent"
@@ -24,101 +22,80 @@
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent">
 
-
         <androidx.constraintlayout.widget.ConstraintLayout
             android:layout_width="match_parent"
-            android:layout_height="match_parent"
-
-
-            >
+            android:layout_height="match_parent">
 
             <ImageView
                 android:id="@+id/logo_login"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
                 android:layout_marginTop="0dp"
-                android:padding="20dp"
                 android:gravity="center"
+                android:padding="20dp"
                 android:src="@drawable/login_moving_logo"
-
                 app:layout_constraintEnd_toEndOf="parent"
                 app:layout_constraintStart_toStartOf="parent"
                 app:layout_constraintTop_toTopOf="parent" />
 
             <TextView
-                android:id="@+id/txt_login_id"
+                android:id="@+id/txt_login_name"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
                 android:layout_gravity="start"
                 android:layout_marginHorizontal="20dp"
-
-
                 android:layout_marginTop="26dp"
-                android:text="@string/emailaddress"
+                android:text="이름"
                 android:textColor="#8CAAB9"
                 android:textSize="18sp"
                 android:textStyle="bold"
                 app:layout_constraintEnd_toEndOf="parent"
                 app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toBottomOf="@id/logo_login"
-
-                />
-
+                app:layout_constraintTop_toBottomOf="@id/logo_login" />
 
             <EditText
-                android:id="@+id/et_login_id"
+                android:id="@+id/et_login_name"
                 android:layout_width="match_parent"
                 android:layout_height="55dp"
-
                 android:layout_marginHorizontal="20dp"
                 android:background="#4D455A64"
-                android:hint="@string/emailaddress"
+                android:hint="@string/inputName"
                 android:inputType="text"
                 android:padding="16dp"
-
-
                 android:textColor="#FFFFFF"
                 app:layout_constraintEnd_toEndOf="parent"
                 app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toBottomOf="@id/txt_login_id"
-
-
-                />
+                app:layout_constraintTop_toBottomOf="@id/txt_login_name" />
 
             <TextView
-                android:id="@+id/txt_login_password"
+                android:id="@+id/txt_login_locale"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
                 android:layout_gravity="start"
                 android:layout_marginHorizontal="20dp"
-
-
                 android:layout_marginTop="26dp"
-                android:text="@string/password"
+                android:text="암호명"
                 android:textColor="#8CAAB9"
                 android:textSize="18sp"
                 android:textStyle="bold"
                 app:layout_constraintEnd_toEndOf="parent"
                 app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toBottomOf="@id/et_login_id"
-
-
-                />
+                app:layout_constraintTop_toBottomOf="@id/et_login_name" />
 
             <EditText
-                android:id="@+id/et_login_password"
+                android:id="@+id/et_login_locale"
                 android:layout_width="match_parent"
                 android:layout_height="55dp"
                 android:layout_marginHorizontal="20dp"
                 android:layout_marginBottom="32dp"
                 android:background="#4D455A64"
-                android:hint="@string/password"
+                android:hint="암호명을 입력해주세요."
                 android:inputType="textPassword"
                 android:padding="16dp"
                 android:textColor="#FFFFFF"
                 app:layout_constraintEnd_toEndOf="parent"
                 app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toBottomOf="@id/txt_login_password" />
+                app:layout_constraintTop_toBottomOf="@id/txt_login_locale" />
 
             <TextView
                 android:id="@+id/buttonLogin"
@@ -127,52 +104,31 @@
                 android:layout_marginHorizontal="20dp"
                 android:layout_marginTop="60dp"
                 android:background="#99FED36A"
-
-
                 android:gravity="center"
-
                 android:text="@string/logIn"
                 android:textSize="20dp"
                 android:textStyle="bold"
                 app:layout_constraintBottom_toTopOf="@id/buttonSignUp"
                 app:layout_constraintEnd_toEndOf="parent"
                 app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toBottomOf="@id/et_login_password"
-
-
-                />
+                app:layout_constraintTop_toBottomOf="@id/et_login_locale" />
 
             <TextView
                 android:id="@+id/buttonSignUp"
                 android:layout_width="match_parent"
                 android:layout_height="55dp"
-                android:textSize="20dp"
-                android:gravity="center"
-
-
                 android:layout_marginHorizontal="20dp"
                 android:layout_marginTop="26dp"
                 android:layout_marginBottom="26dp"
                 android:background="#991F63AA"
-                android:textStyle="bold"
-
+                android:gravity="center"
                 android:text="@string/signUp"
-
-
+                android:textSize="20dp"
+                android:textStyle="bold"
                 app:layout_constraintBottom_toBottomOf="parent"
                 app:layout_constraintEnd_toEndOf="parent"
                 app:layout_constraintStart_toStartOf="parent"
-
-                app:layout_constraintTop_toBottomOf="@+id/buttonLogin"
-
-
-                />
-
-
+                app:layout_constraintTop_toBottomOf="@+id/buttonLogin" />
         </androidx.constraintlayout.widget.ConstraintLayout>
-
-
     </ScrollView>
-
-
 </androidx.constraintlayout.widget.ConstraintLayout>
