+++ app/src/main/res/layout/activity_sign_up.xml
@@ -47,7 +47,7 @@
             android:layout_marginEnd="24dp"
             android:background="#455A64"
             android:ems="10"
-            android:hint="@string/name"
+            android:hint="@string/inputName"
             android:inputType="text"
             android:padding="12dp"
             android:textColor="#FFFFFF"
